+++ hs-bindgen/CHANGELOG.md
@@ -26,6 +26,9 @@
   ([#76](https://github.com/well-typed/hs-bindgen/issues/76)). Export items
   are module-qualified (e.g. `Example.myFunc`) to avoid ambiguity with
   Prelude names.
+* The command line client has a new `internal frontend` sub-command. It takes a
+  frontend pass name (e.g. `internal frontend select`), dumping the result of
+  the provided frontend pass.
 * Add `--post-qualified-imports` flag to generate post-qualified imports
   (`import Data.Proxy qualified`) instead of pre-qualified imports. This adds
   the `ImportQualifiedPost` language extension to generated modules.